requirements.txt
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.8.0
appdirs @ file:///home/conda/feedstock_root/build_artifacts/appdirs_1733753955715/work
beautifulsoup4 @ file:///home/conda/feedstock_root/build_artifacts/beautifulsoup4_1738740337718/work
blinker==1.8.2
Bottleneck @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_55txi4fy1u/croot/bottleneck_1731058642212/work
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1725267563793/work
cachetools==5.5.1
certifi==2024.8.30
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1725560622100/work
charset-normalizer==3.4.0
click==8.1.8
contourpy @ file:///Users/runner/miniforge3/conda-bld/contourpy_1729602398863/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1733332471406/work
distro==1.9.0
duckduckgo_search==7.3.2
easyocr==1.7.2
et_xmlfile==2.0.0
exceptiongroup==1.2.2
filelock==3.16.1
Flask==2.0.1
Flask-Cors==3.0.10
fonttools==4.54.1
frozendict @ file:///Users/runner/miniforge3/conda-bld/frozendict_1728841376066/work
fsspec==2024.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.67.0
greenlet==3.1.1
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
html5lib @ file:///home/conda/feedstock_root/build_artifacts/html5lib_1734075161666/work
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
imageio==2.37.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
kiwisolver @ file:///Users/runner/miniforge3/conda-bld/kiwisolver_1725459226886/work
lazy_loader==0.4
linecache2==1.0.0
lxml @ file:///Users/runner/miniforge3/conda-bld/lxml_1715598702210/work
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
multitasking @ file:///home/conda/feedstock_root/build_artifacts/multitasking_1622493113162/work
munkres==1.1.4
networkx==3.2.1
ninja==1.11.1.3
numexpr @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_b3kvvt6tc6/croot/numexpr_1730215947700/work
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1732314388626/work/dist/numpy-2.0.2-cp39-cp39-macosx_11_0_arm64.whl#sha256=f5ae29dd6f735c6dcad61765cffcfb32661c04e515c759b21399c7a3a39430a4
openai==1.63.0
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==24.1
pandas @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_4aifrweohv/croot/pandas_1732735109535/work/dist/pandas-2.2.3-cp39-cp39-macosx_11_0_arm64.whl#sha256=25b432d8b6b67061b3242c9572f7a6ac74563a28e9b090025e6871d777a433c2
patsy @ file:///home/conda/feedstock_root/build_artifacts/patsy_1733792384640/work
pdf2image==1.17.0
peewee @ file:///Users/runner/miniforge3/conda-bld/peewee_1720385358455/work
pika==1.3.2
pillow @ file:///Users/runner/miniforge3/conda-bld/pillow_1712154530882/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1733232627818/work
playwright==1.50.0
pluggy==1.5.0
primp==0.12.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==6.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
PyJWT==2.3.0
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pyparsing==3.2.0
PyPDF2==3.0.1
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
pytest==8.3.4
pytest-asyncio==0.25.3
python-bidi==0.6.6
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
python-dotenv==0.19.0
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1706886791323/work
PyYAML==6.0.2
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1733217035951/work
rsa==4.9
scikit-image==0.24.0
scipy @ file:///Users/runner/miniforge3/conda-bld/scipy-split_1716470312029/work/dist/scipy-1.13.1-cp39-cp39-macosx_11_0_arm64.whl#sha256=95cc9a6aceed50037865da4a45baac05fd0dcee7e557c6149fde270f2ffebafc
seaborn @ file:///home/conda/feedstock_root/build_artifacts/seaborn-split_1714494649443/work
shapely==2.0.7
six==1.16.0
sniffio==1.3.1
soupsieve @ file:///home/conda/feedstock_root/build_artifacts/soupsieve_1693929250441/work
statsmodels @ file:///Users/runner/miniforge3/conda-bld/statsmodels_1727986732968/work
sympy==1.13.1
thop==0.1.1.post2209072238
tifffile==2024.8.30
tomli==2.2.1
torch==2.5.0
torchvision==0.20.0
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1732615925783/work
tqdm==4.66.5
traceback2==1.4.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1733188668063/work
tzdata==2024.2
ultralytics==8.3.21
ultralytics-thop==2.0.9
unicodedata2 @ file:///Users/runner/miniforge3/conda-bld/unicodedata2_1736692557404/work
unittest2==1.1.0
uritemplate==4.1.1
urllib3==2.2.3
webencodings @ file:///home/conda/feedstock_root/build_artifacts/webencodings_1733236011802/work
Werkzeug==2.0.1
yfinance @ file:///home/conda/feedstock_root/build_artifacts/yfinance_1739938562631/work
zipp==3.20.2
zstandard==0.23.0
